+++ packages/button/src/types.ts
@@ -1,5 +1,5 @@
 import { BaseFontSize } from '@leafygreen-ui/tokens';
-import { HTMLProps } from 'react';
+import React, { HTMLProps } from 'react';
 
 const Variant = {
   Default: 'default',
@@ -98,7 +98,21 @@ interface ButtonProps {
    * Default: `undefined`
    */
   href?: string;
-  as?: keyof JSX.IntrinsicElements;
+
+  /**
+   * The component or HTML Element that the button is rendered as.
+   *
+   * To use with NextJS Links, pass in a component that wraps the Link:
+   * ```js
+   * const Linker = ({ href, children, ...props }) => (
+   *  <NextLink href={href}>
+   *    <a {...props}>{children}</a>
+   *  </NextLink>
+   * );
+   * <Button as={Linker} />
+   * ```
+   */
+  as?: React.ElementType<any>;
 }
 
 export { Variant, Size, ButtonProps };